factor build into its own module
